[Rename] 파일/디렉토리 이름 변경, 불필요한 코드 삭제
