+++ js/App/Main/MainContent/NewsListView/Contents.js
