+++ js/App/Main/MainContent/NewsListView/Contents/Header.js
