+++ js/App/Main/MainContent/NewsListView/Contents/News/SubNews.js
@@ -2,7 +2,7 @@
 리스트 뉴스 컨테이너 컴포넌트
 */
 
-export default function Right($target, props) {
+export default function SubNews($target, props) {
   this.render = () => {
     const $div = document.createElement("div");
     const $ul = document.createElement("ul");
